+++ Sources/ProjectSpec/Settings.swift
@@ -105,3 +105,16 @@ public func += (lhs: inout BuildSettings, rhs: BuildSettings?) {
     guard let rhs = rhs else { return }
     lhs.merge(rhs)
 }
+
+extension Settings: JSONEncodable {
+    public func toJSONValue() -> Any {
+        if groups.count > 0 || configSettings.count > 0 {
+            return [
+                "base": buildSettings,
+                "groups": groups,
+                "configs": configSettings.mapValues { $0.toJSONValue() }
+            ]
+        }
+        return buildSettings
+    }
+}