+++ Sources/XcodeGenKit/SourceGenerator.swift
@@ -224,7 +224,8 @@ class SourceGenerator {
                  "gpx",
                  "lproj",
                  "xcfilelist",
-                 "apns":
+                 "apns",
+                 "pch":
                 return nil
             default:
                 return .resources